Fixed some stuff

+++ include/component/Transform.hpp
@@ -21,8 +21,8 @@ namespace emblem {
 
     struct Transform {
         Direction facing = NORTH;
-        size_t x = 0;
-        size_t y = 0;
+        int64_t x = 0;
+        int64_t y = 0;
 
         void face(Direction direction);
         void face(const Point &point);
